+++ packages/alfa-media/src/feature/index.ts
@@ -1,0 +1,1 @@
+export * from "./feature";
