+++ packages/alfa-media/src/index.ts
@@ -1,2 +1,7 @@
+export * from "./condition/";
+export * from "./feature";
 export * from "./media";
+export * from "./query";
+export * from "./query-list";
 export * from "./resolver";
+export * from "./type";
